use ShortData in MicAudioProvider

+++ src/dataproviders/MicAudioProvider.java
@@ -7,7 +7,7 @@ import javax.sound.sampled.AudioSystem;
 import javax.sound.sampled.DataLine;
 import javax.sound.sampled.TargetDataLine;
 
-import datatypes.DoubleData;
+import datatypes.ShortData;
 
 public class MicAudioProvider extends DataProvider {
     private int sampleRate;
@@ -30,16 +30,16 @@ public class MicAudioProvider extends DataProvider {
             dataline.start();
     
             byte[] buffer = new byte[this.bufferSize];
-            double[] samples = new double[this.bufferSize / 2];
+            short[] samples = new short[this.bufferSize / 2];
 
             ByteBuffer byteBuffer = ByteBuffer.wrap(buffer);
 
             while(audioStream.read(buffer) >= 0) {
                 byteBuffer.position(0);
                 for(int i = 0; i < this.bufferSize/2; i++) {
-                   samples[i] = (double)byteBuffer.getShort();
+                   samples[i] = byteBuffer.getShort();
                 }
-                this.notifyListeners(new DoubleData(samples));
+                this.notifyListeners(new ShortData(samples));
             }
         } catch (Exception e) {
             e.printStackTrace();
